+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 **2.1.0**
 
 - Added support for ``typings.Type`` (available in Python 3.5.2+)
+- Added a third, ``sys.setprofile()`` based type checking approach (``typeguard.TypeChecker``)
 
 **2.0.2** (2016-12-17)